chore: move deployment fixtures
